% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPD.R
\name{CPD}
\alias{CPD}
\title{CPD}
\usage{
CPD(xorigin, ximputed)
}
\arguments{
\item{xorigin}{numeric matrix. The original data. Can not contain missing values.}

\item{ximputed}{numeric matrix. The imputed data. Can not contain missing values.}
}
\value{
numeric
}
\description{
Spearman correlation between pairwise distances in the original data and imputed data. 
CPD quantifies preservation of the global structure after imputation.
}
