% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidenceToMatrix.R
\name{evidenceToMatrix}
\alias{evidenceToMatrix}
\title{Creates intensity matrix from tabular data in evidence table of MaxQuant}
\usage{
evidenceToMatrix(
  evidence,
  run_id = "Raw.file",
  peptide_id = "PeptideID",
  return_MAList = FALSE
)
}
\arguments{
\item{evidence}{data.frame. The evidence table read from evidence.txt, or \code{mspip} output}

\item{run_id}{character. The name of the column of evidence containing the run/raw file name.
These form the columns of the intensity data matrix.}

\item{peptide_id}{character. The name of the column of evidence containing the peptide ids.
These form the rows of the intensity data matrix.}

\item{return_MAList}{logical. If TRUE, returns a \code{MAList} object storing both the
intensity data matrix and observation-level weights from
\code{mspip} (propagation confidence score), otherwise returns a matrix.}
}
\value{
a numeric matrix of intensity data, or a \code{MAList} object containing
such data and observation-level weights from \code{mspip}
}
\description{
Every \code{Modified sequence} - \code{Charge} is considered as a precursor feature.
Only the feature with maximum intensity is retained. The columns are run names, the rows
are peptide ids (in the \code{Modified.sequence_Charge} format)
}
\seealso{
mspip
}
\author{
Soroor Hediyeh-zadeh
}
