% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStructuralMetrics.R
\name{gromov_wasserstein}
\alias{gromov_wasserstein}
\title{Gromov-Wasserstein distance}
\usage{
gromov_wasserstein(x, y, k, min.mean = 0.1)
}
\arguments{
\item{x}{numeric matrix}

\item{y}{numeric matrix}

\item{k}{number of Principal Components used to compute Gromov-Wasserstein distance}

\item{min.mean}{numeric. Minimum mean log-intensity for peptides to be retained}
}
\value{
numeric GW distance
}
\description{
This function is called internally by \code{computeStructuralMetrics}.
Users should not need to call this function directly.
}
\keyword{internal}
