% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msImpute.R
\name{msImpute}
\alias{msImpute}
\title{Peptide-level imputation in mass spectrometry label-free proteomics by low-rank approximation}
\usage{
msImpute(object, rank.max = NULL, lambda = NULL, thresh = 1e-05,
  maxit = 100, trace.it = FALSE, warm.start = NULL,
  final.svd = TRUE)
}
\arguments{
\item{object}{Numeric matrix  where missing values are denoted by NA. Rows are peptides, columns are samples.}

\item{rank.max}{Numeric. This restricts the rank of the solution. is set to min(dim(\code{object})-1) by default.}

\item{lambda}{Numeric. Nuclear-norm regularization parameter. Controls the low-rank property of the solution
to the matrix completion problem. By default, it is determined at the scaling step. If set to zero
the algorithm reverts to "hardImputation", where the convergence will be slower.}

\item{thresh}{Numeric. Convergence threshold. Set to 1e-05, by default.}

\item{maxit}{Numeric. Maximum number of iterations of the algorithm before the algorithm is converged. 100 by default.}

\item{trace.it}{Logical. Prints traces of progress of the algorithm.}

\item{warm.start}{List. A SVD object can be used to initialize the algorithm instead of random initialization.}

\item{final.svd}{Logical. Shall final SVD object be saved? The solutions to the matrix completion problems are computed from U, D and V components of final SVD.}
}
\value{
Missing values are imputed by low-rank approximation of the input matrix. If input is a numeric matrix,
a numeric matrix of identical dimensions is returned. If \code{x} is a \code{MAList} object, the \code{E} component is
replaced with the completed matrix, and the updated \code{MAList} object is returned. Non-NA entries remain unmodified.
}
\description{
Returns a completed peptide intensity matrix where missing values (NAs) are imputated
by low-rank approximation of the input matrix. Non-NA entries remain unmodified. \code{msImpute} requires at least 4
non-missing measurements per peptide across all samples. It is assumed that peptide intensities  (DDA), or MS1/MS2 normalised peak areas (DIA),
are log2-transformed and normalised (e.g. quantile normalisation).
}
\details{
\code{msImpute} operates on the softImpute-ALS algorithm.
For more details on the underlying algorithm, please see \code{\link[softImpute]{softImpute}} package.
}
\examples{
set.seed(101)
n=200
p=100
J=50
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J),n,J)\%*\%matrix(rnorm(J*p),J,p)+matrix(rnorm(np),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA
xna <- scaleData(xna)
xcomplete <- msImpute(object=xna)
}
\seealso{
selectFeatures, scaleData
}
