% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxd007959.R
\docType{data}
\name{pxd007959}
\alias{pxd007959}
\title{Processed peptide intensity matrix and experimental design table from PXD007959 study}
\format{A list of two: samples (data frame of sample descriptions), and y (numeric matrix of peptide intensity values)}
\source{
\url{http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD007959}
}
\usage{
pxd007959
}
\description{
Extracellular vesicles isolated from the descending colon of pediatric patients with inflammatory bowel disease
and control patients. Characterizes the proteomic profile of extracellular vesicles isolated from the descending colon
of pediatric patients with inflammatory bowel disease and control participants. This object contains data from peptide.txt
table output by MaxQuant. Rows are Modified Peptide IDs. Charge state variations are treated as distinct peptide species.
Reverse complements and contaminant peptides are discarded. Peptides with more than 4 observed intensity values are retained.
Additionally, qualified peptides are required to map uniquely to proteins.
Two of the samples with missing group annotation were excluded.
The peptide.txt and experimentalDesignTemplate files can be downloaded as RDS object from \url{https://github.com/soroorh/proteomicscasestudies}.
Code for data processing is provided in package vignette.
}
\references{
Zhang X, Deeke SA, Ning Z, Starr AE, Butcher J, Li J, Mayne J, Cheng K, Liao B, Li L, Singleton R, Mack D, Stintzi A, Figeys D, Metaproteomics reveals associations between microbiome and intestinal extracellular vesicle proteins in pediatric inflammatory bowel disease. Nat Commun, 9(1):2873(2018)
}
\keyword{datasets}
