% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxd010943.R
\docType{data}
\name{pxd010943}
\alias{pxd010943}
\title{SWATH-MS Analysis of Gfi1-mutant bone marrow neutrophils}
\format{
A matrix
}
\source{
\url{http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD010943}
}
\usage{
pxd010943
}
\description{
Contains Peak Area for peptides in PXD010943.
This study investigates the proteomic alterations in bone marrow neutrophils isolated from 5-8 week old Gfi1+/-, Gfi1K403R/-,
Gfi1R412X/-, and Gfi1R412X/R412X mice using the SWATH-MS technique. This dataset consists of 13 SWATH-DIA  runs on a TripleTOF 5600 plus (SCIEX).
Rows are peptides. Charge state variations are treated as distinct peptide species. Peptides with more than 4 observed intensity values are retained.
The peptide.txt and experimentalDesignTemplate files can be downloaded as RDS object from \url{https://github.com/soroorh/proteomicscasestudies}.
Code for data processing is provided in package vignette.
}
\references{
Muench DE, Olsson A, Ferchen K, Pham G, Serafin RA, Chutipongtanate S, Dwivedi P, Song B, Hay S, Chetal K, Trump-Durbin LR, Mookerjee-Basu J, Zhang K, Yu JC, Lutzko C, Myers KC, Nazor KL, Greis KD, Kappes DJ, Way SS, Salomonis N, Grimes HL, Mouse models of neutropenia reveal progenitor-stage-specific defects. Nature, 582(7810):109-114(2020)
}
\keyword{datasets}
