% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxd014777.R
\docType{data}
\name{pxd014777}
\alias{pxd014777}
\title{Processed peptide intensity matrix from PXD014777 study}
\format{
A matrix
}
\source{
\url{http://proteomecentral.proteomexchange.org/cgi/GetDataset?ID=PXD014777}
}
\usage{
pxd014777
}
\description{
A Trapped Ion Mobility Spectrometry (TIMS) dataset of blood plasma from a number of patients acquired in two batches.
This is a technical dataset
published by MaxQuant to benchmark their software for ion mobility enhanced shotgun proteomics.
Rows are Modified Peptide IDs. Charge state variations are treated as distinct peptide species.
For peptides with multiple identification types, the intensity is considered to be the median of reported intensity values.
Reverse complememts and contaminant peptides are discarded.
Peptides with more than 4 observed intensity values are retained.
This object contains data from peptide.txt table output by MaxQuant.
The evidence.txt file can be downloaded as RDS object from \url{https://github.com/soroorh/proteomicscasestudies}.
Code for data processing is provided in package vignette.
}
\references{
Prianichnikov N, Koch H, Koch S, Lubeck M, Heilig R, Brehmer S, Fischer R, Cox J, MaxQuant Software for Ion Mobility Enhanced Shotgun Proteomics. Mol Cell Proteomics, 19(6):1058-1069(2020)
}
\keyword{datasets}
