% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStructuralMetrics.R
\name{computeStructuralMetrics}
\alias{computeStructuralMetrics}
\title{Metrics for the assessment of post-imputation structural preservation}
\usage{
computeStructuralMetrics(x, group)
}
\arguments{
\item{x}{numeric matrix. An imputed data matrix.}

\item{group}{factor. A vector of biological groups, experimental conditions or phenotypes (e.g. control, treatment).}
}
\value{
list of two metrics: withinness (sum of squared distances within a phenotype group on log scale),
betweenness (sum of squared distances between the phenotypes on log scale)
}
\description{
For an imputed dataset, it computes within phenotype/experimental condition similarity (i.e. preservation of local structures)
and between phenotype distances (preservation of global structures). Metrics are on log scale.
}
\details{
For each group of experimental conditions (e.g. treatment and control), the group centroid is calculated as the average
of observations. Withinness for each group is computed as sum of the squared distances between samples in that group and the group centroid. Betweenness is computed as
sum of the squared distances between group centroids. When comparing imputation approaches, the optimal imputation strategy should minimize the within
group distances (hence smaller withinness), and maximizes between group distances, hence larger betweenness.
}
