% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Select features with high biological dropout rate}
\usage{
selectFeatures(object, n_features = 500, suppress_plot = FALSE)
}
\arguments{
\item{object}{Numeric matrix where missing values are denoted by NA.
Rows are peptides, columns are samples.}

\item{n_features}{Numeric, number of features with high dropout rate. 500 by default.}

\item{suppress_plot}{Logical show plot of dropouts vs abundances.}
}
\value{
A data frame with a logical column denoting the selected features
}
\description{
Fits a linear model to peptide dropout rate against peptide abundance. The selected features (peptides) can be
used to determine if data is Missing Not At Random (MNAR). Users should note that \code{msImpute} assumes peptides
are Missing At Random (MAR).
}
\examples{
set.seed(101)
n=800
p=100
J=50
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J),n,J)\%*\%matrix(rnorm(J*p),J,p)+matrix(rnorm(np),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA
rownames(xna) <- 1:nrow(xna)
hdp <- selectFeatures(xna, n_features=500,  suppress_plot=FALSE)
# construct matrix M to capture missing entries
M <- ifelse(is.na(xna),1,0)
M <- M[hdp$msImpute_feature,]
# plot a heatmap of missingness patterns for the selected peptides
library(ComplexHeatmap)
hm <- Heatmap(M,
column_title = "dropout pattern, columns ordered by dropout similarity",
              name = "",
              col = c("#8FBC8F", "#FFEFDB"),
              show_row_names = FALSE,
              show_column_names = TRUE,
              cluster_rows = TRUE,
              cluster_columns = TRUE,
              show_column_dend = FALSE,
              show_row_dend = FALSE,
              row_names_gp =  gpar(fontsize = 7),
              column_names_gp = gpar(fontsize = 8),
              heatmap_legend_param = list(#direction = "horizontal",
              heatmap_legend_side = "bottom",
              labels = c("observed","missing"),
              legend_width = unit(6, "cm")),
         )
hm <- draw(hm, heatmap_legend_side = "left")
}
\seealso{
scaleData, msImpute
}
