% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFeatures.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Select features with high biological dropout rate}
\usage{
selectFeatures(object, n_features = 500, suppress_plot = FALSE)
}
\arguments{
\item{object}{Numeric matrix or \code{MAList} object from \link{limma} where missing values are denoted by NA.
Rows are peptides, columns are samples.}

\item{n_features}{Numeric, number of features with high dropout rate. 500 by default.}

\item{suppress_plot}{Logical show plot of dropouts vs abundances.}
}
\value{
A data frame with a logical column denoting the selected features
}
\description{
Fits a linear model to peptide dropout rate against peptide abundance. The selected features (peptides) can be
used to determine if data is Missing Not At Random (MNAR). Users should note that \code{msImpute} assumes peptides
are Missing At Random (MAR).
}
\examples{
set.seed(101)
n=200
p=100
J=50
np=n*p
missfrac=0.3
x=matrix(rnorm(n*J),n,J)\%*\%matrix(rnorm(J*p),J,p)+matrix(rnorm(np),n,p)/5
ix=seq(np)
imiss=sample(ix,np*missfrac,replace=FALSE)
xna=x
xna[imiss]=NA
selectFeatures(xna, n_features=500,  suppress_plot=FALSE)
}
