% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStructuralMetrics.R
\name{withinness}
\alias{withinness}
\title{Withinness}
\usage{
withinness(x, class_label)
}
\arguments{
\item{x}{numeric matrix}

\item{class_label}{factor. The experimental group of the samples (e.g. control, treatment, WT, Het etc.)}
}
\value{
numeric. Sum of the squared distances of all samples within an experimental group from group centroid
}
\description{
This function is called internally by \code{computeStructuralMetrics}.
Users should not need to call this function directly.
}
\examples{
data(pxd007959)
y <- pxd007959$y
y <- y[complete.cases(y),]
withinness(y, as.factor(pxd007959$samples$group))

}
